pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spyreapi"
version = "0.0.2"
authors = [
  { name="Sanjid Sharaf", email="sanjidsharaf1@gmail.com" },
]
description = "A robust and extensible Python client for interacting with the [Spire Business Software API](https://developer.spiresystems.com/reference). This client provides an object-oriented interface to get, create, update, delete, query, filter, sort, and manage various Spire modules such as Sales Orders, Invoices, Inventory Items, and more."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
  "requests>=2.32.0",
  "httpx>=0.28.0",
  "pydantic>=2.11.0",
  "python-dotenv>=1.1.0",
  "pandas>=2.3.0",
  "openpyxl>=3.1.5",
  "numpy>=2.3.0"
]

[project.urls]
Homepage = "https://github.com/sanjid-sharaf/spyre/"
Issues = "https://github.com/sanjid-sharaf/spyre/issues"